requirements.txt
# torch==1.9.1+cpu
torchvision==0.10.1
torchaudio==0.9.1
https://download.pytorch.org/whl/cpu/torch-1.9.1%2Bcpu-cp39-cp39-linux_x86_64.whl
transformers==3.0.2
numpy==1.19.1
flask
joblib==0.16.0
# sentencepiece==0.1.91
# urllib3==1.25.10